# backend/app/api/static/docs/openapi-3.yaml
openapi: 3.0.0
info:
  description: Track, Manage, and Organize your Things.
  title: HomeBot API
  contact:
    name: HomeBot Team
    url: https://github.com/pritish-codes/homebot
  version: "1.0"
paths:
...
  /v1/group/exports:
    get:
      security:
        - Bearer: []
      description: Returns export job rows for the caller's group, newest first.
      tags:
        - Group
      summary: List Collection Exports
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v1.Results-repo_ExportOut"
    post:
      security:
        - Bearer: []
      description: Creates a pending export row and enqueues the build job. Poll the
        listing endpoint or watch the WebSocket for completion.
      tags:
        - Group
      summary: Start a Collection Export
      responses:
        "202":
          description: Accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/repo.ExportOut"
  "/v1/group/exports/{id}":
    get:
      security:
        - Bearer: []
      tags:
        - Group
      summary: Get an Export
      parameters:
        - description: Export ID
          name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/repo.ExportOut"
    delete:
      security:
        - Bearer: []
      description: Deletes the export row and its blob artifact.
      tags:
        - Group
      summary: Delete an Export
      parameters:
        - description: Export ID
          name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: No Content
  "/v1/group/exports/{id}/download":
    get:
      security:
        - Bearer: []
      tags:
        - Group
      summary: Download an Export Artifact
      parameters:
        - description: Export ID
          name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/zip:
              schema:
                type: string
                format: binary
  /v1/group/import:
    post:
      security:
        - Bearer: []
      description: Uploads a collection-export zip and enqueues the import job. The
        destination group must be empty. Returns the tracked import row so
        clients can poll for progress.
      tags:
        - Group
      summary: Import a Collection Zip
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  description: Export zip
                  type: string
                  format: binary
              required:
                - file
        required: true
      responses:
        "202":
          description: Accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/repo.ExportOut"
...
  "/v1/labelmaker/asset/{id}":
    get:
      security:
        - Bearer: []
      tags:
        - Items
      summary: Get Asset label
      parameters:
        - description: Asset ID
          name: id
          in: path
          required: true
          schema:
            type: string
        - description: Print this label, defaults to false
          name: print
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: image/png
          content:
            application/json:
              schema:
                type: string
  "/v1/labelmaker/item/{id}":
    get:
      security:
        - Bearer: []
      tags:
        - Items
      summary: Get Item label
      parameters:
        - description: Item ID
          name: id
          in: path
          required: true
          schema:
            type: string
        - description: Print this label, defaults to false
          name: print
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: image/png
          content:
            application/json:
              schema:
                type: string
  "/v1/labelmaker/location/{id}":
    get:
      security:
        - Bearer: []
      tags:
        - Locations
      summary: Get Location label
      parameters:
        - description: Location ID
          name: id
          in: path
          required: true
          schema:
            type: string
        - description: Print this label, defaults to false
          name: print
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: image/png
          content:
            application/json:
              schema:
                type: string
...
servers:
  - url: /api
components:
  securitySchemes:
    Bearer:
      description: "\"Type 'Bearer TOKEN' to correctly set the API Key\""
      type: apiKey
      name: Authorization
      in: header
  schemas:
...
    repo.ExportOut:
      type: object
      properties:
        artifactPath:
          type: string
        createdAt:
          type: string
        error:
          type: string
        groupId:
          type: string
        id:
          type: string
        kind:
          description: |-
            Kind is "export" for server-produced backup artifacts, "import" for
            user-uploaded restore zips. The lifecycle fields below behave the
            same for both.
          type: string
        progress:
          type: integer
        sizeBytes:
          type: integer
        status:
          type: string
        updatedAt:
          type: string
...
    v1.Results-repo_ExportOut:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/repo.ExportOut"
...